Can't receive a bid after being depledged
After talking to Mike from International who was very concerned with
the number of times people were pledging and its affect on their
well-being, we decided to prevent people who had been depledged from
being repledged. Because people who do not make grades are not
technically "depledged" this does not count for people who do not make
grades. Also pledges that terminate their own pledgeship are not
technically depledged. Finally, as with all bylaws, we can override
this with a 75% majority.