If you dont' make grades, you stay a pledge
We previously passed a bylaw so that pledges who were depledged could
not receive another bid. But we don't want to prevent pledges who
didn't make grades repledging. We've now separated depledging into
"depledging" and "terminating" which occurs due to academics and does
not count as depledging. If a pledge makes it to initiation but doesn't
make grades, they can continue to be a pledge as long as they are in
good academic standing.